+++ base/process/process_metrics.h
@@ -271,7 +271,7 @@ class BASE_EXPORT ProcessMetrics {
 
 #if BUILDFLAG(IS_LINUX) || BUILDFLAG(IS_CHROMEOS) || BUILDFLAG(IS_ANDROID) || \
     BUILDFLAG(IS_AIX)
-  CPU::CoreType GetCoreType(int core_index);
+  CPU::CoreType GetCoreType(uint32_t core_index);
 #endif  // BUILDFLAG(IS_LINUX) || BUILDFLAG(IS_CHROMEOS) ||
         // BUILDFLAG(IS_ANDROID) || BUILDFLAG(IS_AIX)
 
@@ -434,7 +434,7 @@ BASE_EXPORT int ParseProcStatCPU(StringPiece input);
 // This should be used with care as no synchronization with running threads is
 // done. This is mostly useful to guarantee being single-threaded.
 // Returns 0 on failure.
-BASE_EXPORT int GetNumberOfThreads(ProcessHandle process);
+BASE_EXPORT int64_t GetNumberOfThreads(ProcessHandle process);
 
 // /proc/self/exe refers to the current executable.
 BASE_EXPORT extern const char kProcSelfExe[];
